Create submodules/packages to organize the codebase a bit better
