+++ Scripts/assignments/assignment_model.py
