+++ Scripts/assignments/test_assignment_model.py
